feat: add year input
