+++ resources/assets/js/model/reducer/post/initComment.js
@@ -1,13 +1,7 @@
-export default (function () {
-    const yearOption = [];
-    const clock = new Date();
-    let latestYear = clock.getFullYear() - 1911;
-    if (clock.getMonth() < 7) latestYear--;
-    for (let i = 0; i < 5; ++i) {
-        const yearItem = latestYear - i;
-        yearOption.push({ value: yearItem, text: yearItem });
-    }
+import getYearOption from './function/getYearOption';
 
+export default (function () {
+    const { latestYear, yearOption } = getYearOption;
     return {
         id: -1,
         confirm: 'false',
